Making progress with the drawable model class
The drawable model is an abstract class that will act as a template for all obj models within the program.  The primary goal is to reduce the amount of code in the display method of the code and to shift more of it away into reusable classes which can be carried over into future projects.